Add new mex.py SCons tool, make minor changes to matlab.py.
Added a new SCons tool "mex.py" for making mex extensions. This one takes a slightly
different approach than the other "matlab.py" tool:  It opens a temporary file
in Python and calls Matlab with the "-r" option (so called "batch mode") such
that the interesting variables are written into said temporary file.  Once
Matlab is done, the contents of the file are read.

This method is used by (and taken from) the mlabwrap (mlabwrap.sf.net) projects
"setup.py" distutils script to compile their Python C-extension, so it should
work on Windows, though it still needs to be tested.