matlab.py: make it work under Windows.
Incorporated changes to the matlab.py that allow it to work on Windows an Linux,
at the expense of a slight amount of additional fragility.  I.e., we can't put
the entire Matlab block inside a "with" statement because on Windows a file
can't be opened by more than one process at once (see
http://docs.python.org/library/tempfile.html, section
"tempfile.NamedTemporaryFile").  The SConstruct file has also been modified
accordingly.

Obviously, this whole setup is far from perfect, the best option would probably
be to add some additional builders to SCons (so that one could call, e.g.,
"env.Mex()" to create a Mex extension target), instead of just a tool.